lib: refactor
